there are no standard way to printf time_t
    The time_t datatype is a data type in the ISO C library defined for storing system time values. Such values are returned from the standard time() library function. This type is a typedef defined in the standard header. ISO C defines time_t as an arithmetic type, but does not specify any particular type, range, resolution, or encoding for it. Also unspecified are the meanings of arithmetic operations applied to time values.

    Unix and POSIX-compliant systems implement the time_t type as a signed
      integer (typically 32 or 64 bits wide) which represents the number of seconds since the start of the Unix epoch: midnight UTC of January 1, 1970 (not counting leap seconds). Some systems correctly handle negative time values, while others do not. Systems using a 32-bit time_t type are susceptible to the Year 2038 problem.

+++ src/lang/time/t.c
@@ -8,7 +8,7 @@ main(int argc, char **argv) {
 	_unused_(argv);
 
 	time_t epoch = time(0);
-	printf("epoch seconds: %zu\n", epoch);
+	printf("epoch seconds: %llu\n", (unsigned long long)epoch);
 
 	return 0;
 }
